Dramatic cleanup of joomla.css
Joomla.css features selectors dating back to CiviCRM 2 and CiviCRM's core CSS has changed dramatically since them. This Commit removes all styling, then adds back the selectors that appear to be essential for working on Joomla 3 and 4. 

The goal is not to change the Joomla design (beyond removing one backgrund colour specification) but to inherit more from the main civicrm.css. To this end 'em' units are used to inherit more civicrm.css. The classes are name-spaced so this file could run within other CMS and not impact them.